feat: support composite typography tokens
