+++ packages/design-tokens/web/components/navbar.json5
@@ -2,13 +2,8 @@
     navbar: {
         font: {
             title: {
-                "font-size": { value: '{typography.navbar-title.font-size}' },
-                "line-height": { value: '{typography.navbar-title.line-height}' },
-                "letter-spacing": { value: '{typography.navbar-title.letter-spacing}' },
-                "font-weight": { value: '{typography.navbar-title.font-weight}' },
-                "font-family": { value: '{typography.navbar-title.font-family}' },
-                "text-transform": { value: '{typography.navbar-title.text-transform}' },
-                "font-feature-settings": { value: '{typography.navbar-title.font-feature-settings}' }
+                type: 'font',
+                value: '{typography.navbar-title}'
             }
         },
         light: {
