+++ packages/tokens-builder/transforms/attribute/theme.js
@@ -3,9 +3,7 @@ module.exports = (StyleDictionary) => {
         name: 'kbq-attribute/light',
         type: 'attribute',
         matcher: ({ attributes }) => {
-            return (attributes.category === 'markdown' && attributes.item === 'light')
-                || attributes.type === 'light'
-                || attributes.category === 'light';
+            return (attributes.category === 'markdown' && attributes.item === 'light') || attributes.type === 'light';
         },
         transformer: () => ({ light: true })
     });
@@ -14,9 +12,7 @@ module.exports = (StyleDictionary) => {
         name: 'kbq-attribute/dark',
         type: 'attribute',
         matcher: ({ attributes }) => {
-            return (attributes.category === 'markdown' && attributes.item === 'dark')
-                || attributes.type === 'dark'
-                || attributes.category === 'dark';
+            return (attributes.category === 'markdown' && attributes.item === 'dark') || attributes.type === 'dark';
         },
         transformer: () => ({ dark: true })
     });
